corrected the mess after an unwise merge

+++ cron/instant.py
@@ -10,12 +10,12 @@ class Instant:
     list and one observation dictionary for comparison.
     '''
 
-    def __init__(self, _id, forecasts=[], observations={}):
+    def __init__(self, timeplace, forecasts=[], observations={}):
         
-        self._id = _id
+        self.timeplace = timeplace
         self.casts = forecasts
         self.obs = observations
-        self.as_dict = {'_id': self._id,
+        self.as_dict = {'timeplace': self.timeplace,
                         'forecasts': self.casts,
                         'observations': self.obs
                         }
@@ -49,13 +49,36 @@ class Instant:
         :type collection: string
         '''
 
-        from config import database
-        from db_ops import dbncol
+        import config
+        import db_ops
+        
+        col = db_ops.dbncol(client, collection, config.database)
+        
+        if self.itslegit:
+            col.update_one({'timeplace': self.timeplace},
+                           {'$set': self.as_dict},
+                           upsert=True
+                          )
+        return
 
-        col = dbncol(client, collection, database=database)
-        col.update_one({'_id': self._id}, {'$set': self.as_dict}, upsert=True)
 
+def convert(instants):
+    ### THIS FUNCTION IS UNPROVEN ###
+    ''' Convert a list of instants to instant.Instant objects.
     
+    :param instants: dict of instants
+    
+    :return: dict of instant.Instant objects
+    '''
+    converted = {}
+    for key, value in instants:
+        converted[key] = instant.Instant(
+            value['timeplace'],
+            value['forecasts'],
+            value['observation']
+        )
+    return converted
+        
 def cast_count_all(instants):
     ''' get a tally for the forecast counts per document 
 
@@ -71,9 +94,6 @@ def cast_count_all(instants):
     # the forecasts array. Add to the tally for that count.
     for doc in instants:
         n = len(doc['forecasts'])
-        # Move the legit instants to the permenant database
-        if n >= 40:
-            load_legit(doc)
         if n in collection_cast_counts:
             collection_cast_counts[n] += 1
         else:
@@ -82,6 +102,7 @@ def cast_count_all(instants):
 
 
 def sweep(instants):
+    ### THIS FUNCTION IS DOING NOTHING IF IT REFERS TO ANY 'instant' KEY ###
     ''' Move any instant that has a ref_time less than the current next
     ref_time and with self.count less than 40. This is getting rid of the
     instants that are not and will not ever be legit.
@@ -98,8 +119,8 @@ def sweep(instants):
     import config
     import db_ops
     
-    client = MongoClient('localhost', 27017)
-    col = db_ops.dbncol(client, 'instant_temp', config.database)
+    client = MongoClient(config.host, config.port)
+    col = db_ops.dbncol(client, config.instants_collection, config.database)
     n = 0
     # Check the instant type- it could be a dict if it came from the database,
     # or it could be a list if it's a bunch of instant objects, or a pymongo
@@ -123,63 +144,80 @@ def sweep(instants):
         print(f'You want me to sweep instants that are {type(instants)}\'s.')
     return
 
-def find_legit(instants):
-    ### THIS DOES NOT WORK ###
+def find_legit(instants, and_load=True):
     ''' find the 'legit' instants within the list
 
      :param instants: all the instants pulled from the database
      :type instants: list
      :return: list of instants with a complete forecasts array
      '''
-
-    i = [item for item in instants if len(item['forecasts']) >= 40]
-    return f'legit list is {len(i)} items long'
-    ### maybe you should make the instant documents pulled form the database
-    ### represented as Instants in memory. Then you could use the Instant
-    ### methods you've been writing.
-
+    
+    legit_dict = {}
+    legit_load_list = []
+    instants = convert(instants)
+    # Make a load list out of the Instants
+    if and_load:
+        for key, value in instants:
+            if value.itslegit:
+                legit_load_list.append(update_command_for(value.as_dict))
+        print(f'Got the legit_load_list and it is this long: {len(legit_load_list)}')
+        if len(legit_load_list) == 0:
+            pass
+        load_legit(legit_load_list)
+    # Make a lisf of legit instants and return it.
+    for key, value in instants:
+        if value.itslegit:
+            legit_dict[value.timeplace] = value.as_dict
+    return legit_dict
 
 def load_legit(legit):
     ''' Load the 'legit' instants to the remote database and delete from temp.
     This process does not delete the documents upon insertion, but rather holds
-    it until the next time, tries to load it then, and finally, when getting a duplicate
-    key error, deletes the document from its temporary location.
+    it until the next time, tries to load it then, and finally, when getting a
+    duplicate key error, deletes the document from its temporary location.
 
-    :param legit: a single document
-    :type legit: dict
+    :param legit: a single document or a list of update commands
+    :type legit: dict or list
     '''
 
     from pymongo import MongoClient
-    from pymongo.errors import DuplicateKeyError
+    from pymongo.errors import DuplicateKeyError, InvalidOperation
 
     import config
     import db_ops
     
-    client = MongoClient('localhost', 27017)
-    remote_col = db_ops.dbncol(db_ops.Client(config.uri),'legit_inst', config.database)
-    col = db_ops.dbncol(client, 'instant_temp', config.database)
-    # col = dbncol(client, 'legit_inst', config.database)
-    try:
-        check = remote_col.insert_one(legit)
-        if not check:
-            print('was not able to insert to remote db')
-    except DuplicateKeyError:
-        col.delete_one(legit)
-#     ### saved for later, when doing it on bulk ###
-#     col.insert_many(legit_list)
-#     # Now go to the temp_instants collection and delete the instants just
-#     # loaded to legit_inst.
-#     col.delete_many(legit_list)
-#     ### saved for later, when doing it on bulk ###
+    client = MongoClient(config.host, config.port)
+    remote_col = db_ops.dbncol(db_ops.Client(config.uri),
+                               'legit_inst',
+                               config.database
+                              )
+    col = db_ops.dbncol(client,
+                        config.instants_collection,
+                        config.database
+                       )
+    if not isinstance(legit, dict):
+        try:
+            remote_col.bulk_write(legit)
+        except InvalidOperation as e:
+            print(e)
+            client.close()
+            return -1
+    else:
+        try:
+            remote_col.insert_one(legit)
+            print(f'from {type(legit)}, loaded legit')
+        except DuplicateKeyError:
+            col.delete_one(legit)
+    client.close()
     return
 
-
-import time
 
 if __name__ == '__main__':
     ''' Connect to the database, then move all the legit instants to the remote
     database and clear out any instants that are past and not legit.
     '''
+    
+    import time
     
     import config
     import db_ops
@@ -187,10 +225,11 @@ if __name__ == '__main__':
     start_time = time.time() # This is to get the total runtime if this script is
                              # run as __main__
     print('Database sweep in progress...')
-    collection = 'instant_temp'
+    collection = config.instants_collection
     col = db_ops.dbncol(db_ops.Client(config.uri),
-                        collection,
+                        config.instants_collection,
                         config.database)
-    cast_count_all(col.find({}))
-    sweep(col.find({}))
+    instants = db_ops.read_to_dict(col)
+    find_legit(instants, and_load=True)
+    sweep(col.find({}).batch_size(100))
     print(f'Total sweep time was {time.time()-start_time} seconds')
